warehouse/oso_dagster/pyproject.toml
[project]
name = "oso-dagster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "arrow<2.0.0,>=1.3.0",
    "bokeh<4.0.0,>=3.6.1",
    "clickhouse-connect<1.0.0,>=0.7.16",
    "dagster<2.0.0,>=1.8.6",
    "dagster-embedded-elt<1.0.0,>=0.24.0",
    "dagster-gcp<1.0.0,>=0.24.0",
    "dagster-polars<1.0.0,>=0.24.0",
    "dagster-postgres<1.0.0,>=0.24.0",
    "dagster-sqlmesh>=0.22.0",
    "dagster-webserver<2.0.0,>=1.7.16",
    "discord-webhook<2.0.0,>=1.3.1",
    "dlt<2.0.0,>=1.3.0",
    "gcsfs>=2024.6.1,<2026.0.0",
    "google-api-python-client<3.0.0,>=2.116.0",
    "google-cloud-bigquery<4.0.0,>=3.17.1",
    "google-cloud-bigquery-datatransfer<4.0.0,>=3.15.3",
    "google-cloud-resource-manager<2.0.0,>=1.12.3",
    "google-cloud-secret-manager<3.0.0,>=2.20.0",
    "google-cloud-service-management<2.0.0,>=1.8.3",
    "google-cloud-service-usage<2.0.0,>=1.10.3",
    "google-cloud-storage>=2.14.0,<4.0.0",
    "gcloud-aio-storage<10.0.0,>=9.3.0",
    "oss-directory==0.2.5",
    "pendulum<4.0.0,>=3.0.0",
    "polars<2.0.0,>=1.31.0",
    "python-dotenv<2.0.0,>=1.0.1",
    "requests<3.0.0,>=2.31.0",
    "ruamel-yaml<1.0.0,>=0.18.6",
    "sqlalchemy<3.0.0,>=2.0.25",
    "redis>=5.0.7,<7.0.0",
    "githubkit>=0.12.13,<1.0.0",
    "dagster-duckdb-polars<1.0.0,>=0.24.0",
    "google-cloud-bigquery-storage<3.0.0,>=2.25.0",
    "google-auth<3.0.0,>=2.34.0",
    "dagster-k8s<1.0.0,>=0.24.6",
    "pyiceberg[hive,s3fs]<1.0.0,>=0.8.1",
    "pyee>=12.1.1,<14.0.0",
    "aiotrino<1.0.0,>=0.2.3",
    "websockets>=14.1,<16.0",
    "pydantic-settings<3.0.0,>=2.7.0",
    "oso-core",
    "pyoso",
    "cryo>=0.3.2",
    "web3>=7.13.0",
    "pynessie>=0.67.0",
    "scikit-learn>=1.7.2",
    "sqlmesh>=0.227.1",
    "sqlglot>=27.28.1",
    "pandas-gbq>=0.29.2",
    "pandas-stubs>=2.3.2.250926",
    "pandas>=2.3.3",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.24.0,<2.0.0",
    "dagster-webserver<2.0.0,>=1.7.2",
]

[tool.uv.sources]
oso-core = { workspace = true }
pyoso = { workspace = true }
